Various changes
Discovered that CustomEvent data loses its type and so must be
'reconstituted' on the other side. So, used TodoVO.toJson() in the
TodoForm, when creating a new TodoVO and dispatching it in an event.
Then, in the TodoCommand, using TodoVO.fromString() to recreate the
TodoVO originally sent in the event. Here I discovered that
JsonObject.fromString() doesn't appear to work properly.